Moving Playfair code from old to new project (incl tests).
