+++ SimpleCryptoLib/Ciphers/Playfair Cipher/Digraphs/IDigrathGenerator.cs
@@ -1,0 +1,20 @@
+using System.Collections.Generic;
+
+namespace SimpleCryptoLib.Ciphers.Playfair_Cipher.Digraphs;
+
+public interface IDigrathGenerator
+{
+    /// <summary>
+    /// Generates a collection of digraths from specified plain text.
+    /// </summary>
+    /// <param name="plainText">Plain text message.</param>
+    /// <returns>Digraths.</returns>
+    IEnumerable<Digraph> GetMessageDigraths(string plainText);
+
+    /// <summary>
+    /// Generates a collection of digraths from specified, valid cipher text.
+    /// </summary>
+    /// <param name="cipherText">Valid cipher text.</param>
+    /// <returns>Digraths.</returns>
+    IEnumerable<Digraph> GetCipherTextDigraphs(string cipherText);
+}
