+++ SimpleCryptoLib/Ciphers/Playfair Cipher/Digraphs/NullableDigrath.cs
@@ -1,0 +1,26 @@
+namespace SimpleCryptoLib.Ciphers.Playfair_Cipher.Digraphs;
+
+public class NullableDigrath
+{
+    /// <summary>
+    /// First character within the digrath.
+    /// </summary>
+    public char? CharacterOne { get; set; } = null;
+
+    /// <summary>
+    /// Second character within the digrath.
+    /// </summary>
+    public char? CharacterTwo { get; set; } = null;
+
+    public override string ToString()
+    {
+        if (IsComplete()) { return $"'{CharacterOne}{CharacterTwo}'"; }
+
+        return CharacterOne != null ? $"'{CharacterOne}_'" : $"'_{CharacterTwo}'";
+    }
+        
+    private bool IsComplete()
+    {
+        return CharacterOne != null && CharacterTwo != null;
+    }
+}
