+++ SimpleCryptoLib/Ciphers/Playfair Cipher/IPlayfairCipher.cs
@@ -1,0 +1,8 @@
+using SimpleCryptoLib.Ciphers.Common;
+using SimpleCryptoLib.Ciphers.Playfair_Cipher.Key_Management;
+
+namespace SimpleCryptoLib.Ciphers.Playfair_Cipher;
+
+public interface IPlayfairCipher : ICommonCipher<PlayfairKey>
+{
+}
