+++ SimpleCryptoLib/Ciphers/Playfair Cipher/Key Management/IPlayfairKeyManagement.cs
@@ -1,0 +1,10 @@
+using SimpleCryptoLib.Ciphers.Common.Key_Management;
+
+namespace SimpleCryptoLib.Ciphers.Playfair_Cipher.Key_Management;
+
+/// <summary>
+/// Playfair key generation and validation.
+/// </summary>
+public interface IPlayfairKeyManagement : ICipherKeyManagement<PlayfairKey>
+{
+}
