--- SimpleCryptography/Ciphers/Playfair Cipher/Key Management/IPlayfairKeyManagement.cs
@@ -1,11 +1,0 @@
-﻿using SimpleCryptography.Ciphers.Generic.Key_Management;
-
-namespace SimpleCryptography.Ciphers.Playfair_Cipher.Key_Management
-{
-    /// <summary>
-    /// Playfair key generation and validation.
-    /// </summary>
-    public interface IPlayfairKeyManagement : ICipherKeyManagement<PlayfairKey>
-    {
-    }
-}
