+++ SimpleCryptographyUnitTests/SimpleCryptographyUnitTests.csproj
@@ -47,10 +47,6 @@
   <ItemGroup>
     <Compile Include="Cipher Cracker Tests\Frequency Analysis Tests\CharacterFrequencyAnalyzerTests.cs" />
     <Compile Include="Cipher Tests\Monoalphabetic Substitution Cipher\MonoalphabeticSubstitutionCipherTests.cs" />
-    <Compile Include="Cipher Tests\Playfair Cipher\DigramGeneratorTests.cs" />
-    <Compile Include="Cipher Tests\Playfair Cipher\KeyManagementTests.cs" />
-    <Compile Include="Cipher Tests\Playfair Cipher\PlayfairCipherTests.cs" />
-    <Compile Include="Cipher Tests\Playfair Cipher\PlayfairUtilTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
